+++ admin/main.php
@@ -199,8 +199,8 @@ $smarty->registerPlugin('function', 'controlpanel', 'showcontrolpanel');
 
 $v = $lang ['admin'] [$panel] [$action];
 
-$$smarty->assign('panelstrings', $v);
-$$smarty->assign('plang', $v);
+$smarty->assign('panelstrings', $v);
+$smarty->assign('plang', $v);
 
 if (isset($_GET ['mod'])) {