+++ fp-plugins/gdprvideoembed/lang/lang.ja-jp.php
@@ -1,0 +1,42 @@
+<?php
+$lang ['plugin'] ['gdprvideoembed'] = array(
+
+	// for translated content in script file
+	'head_youtube' => 'YouTubeビデオの埋め込み',
+
+	'hint_youtube' => '<p><b>注意:</b> この埋め込みビデオはYouTube, LLC, 901 Cherry Ave, San Bruno, CA 94066, USAによって提供されています。<br>' .//
+		'再生中にYouTubeサーバーへの接続が確立されます。これにより、YouTubeにどのページにアクセスしているかが通知されます。 ' . //
+		'YouTubeアカウントにログインしている場合、YouTubeはあなたの閲覧行動をあなた個人に割り当てることができます。 ' . //
+		'事前にYouTubeアカウントからログアウトしておけば、これを防ぐことができる。</p>' . //
+		'<p>YouTubeの動画が開始されると、プロバイダはユーザーの行動に関する情報を収集するクッキーを使用します。</p>' . //
+		'<p>Google Ads プログラムのクッキーの保存を無効にしている場合、YouTube の動画を視聴する際にそのようなクッキーを期待する必要はありません。 ' . //
+		'ただし、YouTubeは、個人を特定しない利用情報も他のクッキーに保存します。 ' . //
+		'これを防ぐには、ブラウザでクッキーの保存をブロックする必要があります。</p>' . //
+		'<p>YouTube」でのデータ保護に関する詳細は、以下のプロバイダーのプライバシーポリシーに記載されています。: ' . //
+		'<a href="https://policies.google.com/privacy?hl=ja&gl=jp" title="こちらへ https://policies.google.com/privacy?hl=ja&gl=jp" rel="noopener" target="_blank">https://policies.google.com/privacy?hl=ja&gl=jp</a></p>',
+	'link_title_youtube' => 'Video auf YouTube ansehen',
+
+	'head_vimeo' => '埋め込みVimeoビデオ',
+	'hint_vimeo' => '<p><b>注:</b> この埋め込みビデオは、Vimeo, Inc, 555 West 18th Street, New York, New York 10011, USAにより提供されています。<br>' . //
+		'再生中にVimeoサーバーへの接続が確立されます。これにより、Vimeoにどのページにアクセスしているかが伝えられます。 ' . //
+		'お客様がVimeoアカウントにログインしている場合、Vimeoはお客様の閲覧行動をお客様個人に割り当てることができます。 ' . //
+		'事前にVimeoアカウントからログアウトすることでこれを防ぐことができます。</p>' . //
+		'<p>Vimeoビデオが開始されると、プロバイダはユーザーの行動に関する情報を収集するクッキーを使用します。</p>' . //
+		'<p>Vimeo」でのデータ保護に関する詳細は、以下のプロバイダーのプライバシーポリシーに記載されています。: ' . //
+		'<a href="https://vimeo.com/privacy" title="こちらへ https://vimeo.com/privacy" rel="noopener" target="_blank">https://vimeo.com/privacy</a></p>',
+	'link_title_vimeo' => 'Vimeoでビデオを見る',
+
+	'link' => 'ビデオへのリンク',
+	'button_title' => 'このページでビデオを見る',
+	'button' => 'ビデオ再生',
+
+	// for translated content in css file
+	'noscript_1' => 'このビデオはJavaScript',
+	'noscript_2' => 'を有効にしない',
+	'noscript_3' => 'と再生できません。'
+);
+
+$lang ['plugin'] ['gdprvideoembed'] ['errors'] = array (
+	-1 => 'このプラグインを使用するには、BBcodeプラグインがアクティブでなければなりません。'
+);
+?>